requirements.txt
django
Scrapy
redis
celery
django-rest-framework
scrapy-djangoitem